Put back the CardFilter interface just so I can check something out, because npm run build is failing on git side
