+++ src/api/cards.ts
@@ -26,11 +26,3 @@ export const unblockCard = async (client: Axios, cardNumber: string) =>
 
 export const deactivateCard = async (client: Axios, cardNumber: string) =>
   client.put<void>(`cards/deactivate/${cardNumber}`);
-
-export interface CardFilter {
-  cardNumber: string;
-  firstName: string;
-  lastName: string;
-  email: string;
-  cardStatus: string;
-}
